requirements.txt
httpx>=0.27.0
cryptography>=42.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
rich>=13.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
flask>=3.0.0
